config: disable Advanced Protection
Note: Unclear whether the flag overrides are working for this

The notification that asks the user to enable "additional Advanced Protection" features (device
protection) should be disabled by this flag. There were crashes related to GMS attempting to post
this notification if a user with a Google account enrolled in the Advanced Protection program
updates GMS but doesn't reboot their device. This is because the notification needs an Intent to
launch the new Advanced Protection activity, but this activity is disabled in the GMS
AndroidManifest by default. The code that enables / disables the activity appears to only run on
receiving a boot-related broadcast, so not rebooting will result in the activity remaining disabled.

As a result, the Intent for the activity can't be resolved without rebooting. GMS code will use null
for the Intent variable if it can't resolve the activity, but GMS doesn't do a null check for the
variable and encounters an NPE. It's likely this GMS bug never occurs on the stock OS if GMS is
updated inside of an stock OS update to Android 16, as the boot broadcast will be immediately
received by the new GMS code to enable the activity.

Disabling this flag will also make it so that after a reboot, the button for opening the Advanced
Protection activity will no longer appear in Settings -> Apps -> Sandboxed Google Play -> Google
Settings -> All services -> Personal & device safety. Unsure if this works reliably; it seems to
come back on another reboot.

Note that the Advanced Protection activity also has a button to open a webpage to enroll in Advanced
Protection for the Google account, so disabling the entire activity will also remove that flow.
There isn't a flag that allows us to just disable / hide the device protection switch. The entire
activity is also just disabled if the device is not >= Android 16 (i.e. when the new
AdvancedProtectionService is missing in frameworks/base), so they don't seem to intend to allow
just the account enrollment from this activity.